// sealos/frontend/providers/dbprovider/src/types/db.d.ts
export type DBType = `${DBTypeEnum}`;
// ...
export interface DBEditType {
  dbType: DBType;
  dbVersion: string;
  dbName: string;
  replicas: number;
  cpu: number;
  memory: number;
  storage: number;
  labels: { [key: string]: string };
  terminationPolicy: KubeBlockClusterTerminationPolicy;
  autoBackup?: AutoBackupFormType;
  dataSourceId?: number;
}